baselines/DCGen/requirements.txt
selenium==4.18.1
pillow==10.2.0
tqdm==4.66.2
pandas==2.2.1
pillow-avif-plugin==1.4.3
matplotlib>=3.3.4
openai==1.13.3
ipywidgets
ipykernel
Levenshtein
scikit-image>=0.18.1
google-generativeai
pyqt5
lxml==5.2.2
BeautifulSoup4 
requests
flask==3.0.3
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.19.0
opencv-python>=4.5.1
scipy>=1.6.0